add product feature now works

+++ client/pages/api/products/[id].js
@@ -6,7 +6,10 @@ export default async function handler(req, res) {
     const {
         method,
         query: { id },
+        cookies
     } = req;
+
+    const token = cookies.token;
 
     dbConnect()
 
@@ -20,10 +23,24 @@ export default async function handler(req, res) {
     }
 
     if (method === "PUT") {
-       
+        if (!token || token !== process.env.TOKEN) {
+            return res.status(401).json("Not authenticated!")
+        }
+        try {
+            const product = await Product.findByIdAndUpdate(id, req.body, {
+                new: true,
+            });
+            res.status(200).json(product);
+        } catch (error) {
+            res.status(500).json(error);
+        }
+
     }
 
     if (method === "DELETE") {
+        if (!token || token !== process.env.TOKEN) {
+            return res.status(401).json("Not authenticated!")
+        }
         try {
             await Product.findByIdAndDelete(id);
             res.status(201).json("the product has been deleted");
